refactoring core tests
