+++ src/gridworks_test/__init__.py
@@ -1,13 +1,18 @@
+from gridworks_test.stub_actors import GNodeStubRecorder
 from gridworks_test.stub_actors import SupervisorStubRecorder
 from gridworks_test.stub_actors import TimeCoordinatorStubRecorder
+from gridworks_test.stub_actors import load_rabbit_exchange_bindings
 from gridworks_test.wait import AwaitablePredicate
 from gridworks_test.wait import ErrorStringFunction
 from gridworks_test.wait import Predicate
 from gridworks_test.wait import StopWatch
 from gridworks_test.wait import await_for
+from gridworks_test.wait import wait_for
 
 
 __all__ = [
+    "load_rabbit_exchange_bindings",
+    "GNodeStubRecorder",
     "SupervisorStubRecorder",
     "TimeCoordinatorStubRecorder",
     "AwaitablePredicate",
@@ -15,4 +20,5 @@ __all__ = [
     "Predicate",
     "StopWatch",
     "await_for",
+    "wait_for",
 ]
