+++ src/gridworks_test/wait.py
@@ -105,3 +105,35 @@ async def await_for(
             raise ValueError(err_str)
         else:
             return False
+
+
+def wait_for(
+    f: Callable[[], bool],
+    timeout: float,
+    tag: str = "",
+    raise_timeout: bool = True,
+    retry_duration: float = 0.1,
+) -> bool:
+    """Call function f() until it returns True or a timeout is reached. For async tests use await await_for() instead.
+    retry_duration specified the sleep time between calls. If the timeout is reached before f return True, the function
+    will either raise a ValueError (the default), or, if raise_timeout==False, it will return False. Function f is
+    guaranteed to be called at least once. If an exception is raised the tag string will be attached to its message.
+    """
+    now = time.time()
+    until = now + timeout
+    if now >= until:
+        if f():
+            return True
+    while now < until:
+        if f():
+            return True
+        now = time.time()
+        if now < until:
+            time.sleep(min(retry_duration, until - now))
+            now = time.time()
+    if raise_timeout:
+        raise ValueError(
+            f"ERROR. Function {f} timed out after {timeout} seconds. {tag}"
+        )
+    else:
+        return False
